refactor: move custom types to its own file
add a checkResult priority queue, kind of like a std::priority_queue of std::pair<int,std::string> , but using CheckResult instead of pair.

this is used to add many subcheck results and then pick the most important i.e the check with highest priority
