+++ pkg/checkhttp/check.go
@@ -219,53 +219,6 @@ func printVersion(output io.Writer) {
 		runtime.Version())
 }
 
-type capWriter struct {
-	buffer    []byte
-	Cap       uint64
-	size      uint64
-	NoDiscard bool
-}
-
-func (w *capWriter) Write(data []byte) (int, error) {
-	w.size += uint64(len(data))
-	if w.size > w.Cap && w.NoDiscard {
-		return 0, errors.New("could not write body buffer. buffer is full")
-	}
-
-	if w.size > w.Cap {
-		q := w.Cap - uint64(len(w.buffer))
-		if q != 0 {
-			w.buffer = append(w.buffer, data[0:q-1]...)
-		}
-	} else {
-		w.buffer = append(w.buffer, data...)
-	}
-
-	return len(data), nil
-}
-
-func (w *capWriter) Size() uint64 {
-	return w.size
-}
-
-func (w *capWriter) Bytes() []byte {
-	return w.buffer
-}
-
-//nolint:errname // The original author used it as an error type extensively
-type CheckResult struct {
-	msg  string
-	code int
-}
-
-func (e *CheckResult) Error() string {
-	return e.msg
-}
-
-func (e *CheckResult) Code() int {
-	return e.code
-}
-
 type RequestMetadata struct {
 	req            *http.Request
 	res            *http.Response
